chore(provider): rename folder to queryProvider to QueryProvider
- rename file name
- update import paths
